+++ src/app/[lng]/(main)/storylab/(editor)/[slug]/[storyStepId]/StepCalendarModal.tsx
@@ -8,7 +8,7 @@ import React from 'react'
 import { Button } from '@/src/components/Elements/Button'
 import { toast } from '@/src/lib/toast'
 import useStep from '@/src/lib/api/step/useStep'
-import { Calendar } from '@/src/components/Timeline/Calendar'
+import { Calendar } from '@/src/components/Timeline/DatePicker/Calendar'
 import { useForm } from 'react-hook-form'
 
 type Props = {
